+++ Unity Studio/Unity Classes/Shader.cs
@@ -33,7 +33,7 @@ namespace Unity_Studio
                 {
                     if (m_Name != "") { preloadData.Text = m_Name; }
                     else { preloadData.Text = preloadData.TypeString + " #" + preloadData.uniqueID; }
-                    preloadData.SubItems.AddRange(new string[] { preloadData.TypeString, preloadData.Size.ToString() });
+                    preloadData.SubItems.AddRange(new[] { preloadData.TypeString, preloadData.Size.ToString() });
                 }
             }
             else
@@ -56,7 +56,7 @@ namespace Unity_Studio
 
                     if (m_Name != "") { preloadData.Text = m_Name; }
                     else { preloadData.Text = preloadData.TypeString + " #" + preloadData.uniqueID; }
-                    preloadData.SubItems.AddRange(new string[] { preloadData.TypeString, preloadData.Size.ToString() });
+                    preloadData.SubItems.AddRange(new[] { preloadData.TypeString, preloadData.Size.ToString() });
                 }
                 a_Stream.AlignStream(4);
                 m_PathName = a_Stream.ReadAlignedString(a_Stream.ReadInt32());
