+++ Unity Studio/Unity Studio Classes/Lz4DecoderStream.cs
